Moved schedule classes in client-service
