+++ openframe-client-core/src/main/java/com/openframe/client/service/rmm/DeviceOnlineScheduleTriggerService.java
@@ -1,4 +1,4 @@
-package com.openframe.management.service;
+package com.openframe.client.service.rmm;
 
 import com.openframe.data.document.rmm.ScriptSchedule;
 import com.openframe.data.document.rmm.ScriptScheduleMachineAssigned;
