+++ openframe-client-core/src/test/java/com/openframe/client/service/DeviceOnlineScheduleTriggerServiceTest.java
@@ -1,5 +1,7 @@
-package com.openframe.management.service;
+package com.openframe.client.service;
 
+import com.openframe.client.service.rmm.DeviceOnlineScheduleTriggerService;
+import com.openframe.client.service.rmm.ScheduleFireDispatcher;
 import com.openframe.data.document.rmm.ScriptSchedule;
 import com.openframe.data.document.rmm.ScriptScheduleMachineAssigned;
 import com.openframe.data.document.rmm.ScriptScheduleTrigger;
