+++ openframe-client-core/src/test/java/com/openframe/client/service/ScriptExecutionWatchdogServiceTest.java
@@ -1,5 +1,6 @@
-package com.openframe.management.service;
+package com.openframe.client.service;
 
+import com.openframe.client.service.rmm.ScriptExecutionWatchdogService;
 import com.openframe.data.document.rmm.ScriptExecution;
 import com.openframe.data.document.rmm.ExecutionStatus;
 import com.openframe.data.repository.rmm.ScriptExecutionRepository;
