+++ openframe-client-core/src/test/java/com/openframe/client/service/ScriptScheduleExecutionServiceTest.java
@@ -1,5 +1,7 @@
-package com.openframe.management.service;
+package com.openframe.client.service;
 
+import com.openframe.client.service.rmm.ScheduleFireDispatcher;
+import com.openframe.client.service.rmm.ScriptScheduleExecutionService;
 import com.openframe.data.document.rmm.ScriptSchedule;
 import com.openframe.data.document.rmm.ScriptStatus;
 import com.openframe.data.repository.rmm.ScriptScheduleRepository;
